requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
appnope==0.1.0
APScheduler==3.7.0
astroid==2.4.2
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.5.4
base58==2.0.1
base91==1.0.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
cachetools==4.1.1
capstone==4.0.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
ciphey==5.11.1
cipheycore==0.3.2
cipheydists==0.3.34
click==7.1.2
click-spinner==0.1.10
colorama==0.4.3
commonmark==0.9.1
cryptography==3.3.1
Cython==0.29.21
daemonize==2.5.0
dateparser==1.0.0
discord.py==1.6.0
distlib==0.3.1
filelock==3.0.12
google-api-core==1.22.2
google-auth==1.21.2
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
greenlet==0.4.17
idna==2.8
inotify==0.2.10
intervaltree==3.1.0
isort==5.4.2
jedi==0.18.0
lazy-object-proxy==1.4.3
loguru==0.5.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
multidict==4.7.4
nose==1.3.7
numpy==1.19.4
packaging==20.8
pandas==1.1.4
paramiko==2.7.2
parso==0.8.1
Pillow==8.1.0
pipenv==2020.8.13
protobuf==3.13.0
psutil==5.8.0
pwntools==4.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase62==0.4.3
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.9.8
pyelftools==0.27
pyflakes==2.3.0
Pygments==2.7.1
pylint==2.6.0
PyNaCl==1.4.0
pynvim==0.4.2
PyOpenGL==3.1.5
pyparsing==2.4.7
pyperclip==1.8.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
reedsolo==1.5.4
regex==2020.11.13
requests==2.25.1
requests-cache==0.5.2
rich==6.2.0
ROPGadget==6.5
rsa==4.6
sagemath==1.3.0
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.2
toml==0.10.1
typing-extensions==3.7.4.3
tzlocal==2.1
unicorn==1.0.2rc3
urllib3==1.25.10
wrapt==1.12.1
yarl==1.3.0
yaspin==1.0.0
